refactor: fix DRY violations and code quality issues
- Extract parseContextualReferences to shared utils to avoid duplication
- Remove duplicate ContextualItem import
- Clean up unused comments and dead code
- Improve code organization and maintainability
